Throws to manage return frames was a bit excessive. This            wasted lots of time in atomic calls.
git-svn-id: svn://falconpl.org/falcon/core/trunk@1314 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ engine/symbol.cpp
@@ -324,7 +324,7 @@ bool FuncDef::save( Stream *out ) const
          bool res = out->write( ecode, m_codeSize );
          memFree( ecode );
       #else
-         bool res = out->write( m_code, m_codeSize ) == m_codeSize;
+         bool res = out->write( m_code, m_codeSize ) == (int) m_codeSize;
       #endif
 
       if ( ! res )
